Extract AST parser’s S-expression–reading methods into helper module
We’d like to start using these in the preprocessor too.
